Move everything under the DevOptional namespace
