+++ Directory.Build.props
@@ -1,17 +1,17 @@
 <Project>
 
   <PropertyGroup>
-    <AssemblyName>bradselw.SystemResources.$(MSBuildProjectName)</AssemblyName>
-    <RootNamespace>bradselw.SystemResources.$(MSBuildProjectName.Replace(" ", "_"))</RootNamespace>
+    <AssemblyName>DevOptional.System.Resources.$(MSBuildProjectName)</AssemblyName>
+    <RootNamespace>DevOptional.System.Resources.$(MSBuildProjectName.Replace(" ", "_"))</RootNamespace>
   </PropertyGroup>
 
   <PropertyGroup>
     <Authors>bradselw</Authors>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/bradselw/SystemResources</PackageProjectUrl>
+    <PackageProjectUrl>https://github.com/DevOptional/System.Resources</PackageProjectUrl>
     <RepositoryType>git</RepositoryType>
-    <RepositoryUrl>https://github.com/bradselw/SystemResources</RepositoryUrl>
-    <PackageTags>system;resources</PackageTags>
+    <RepositoryUrl>https://github.com/DevOptional/System.Resources</RepositoryUrl>
+    <PackageTags>devops;system;resources</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
